Fixed ESLint import issues
The VS Code ESLint extension was defaulting to Node v14.16.0 instead of
the desired v17.2.0 as defined in the .nvmrc file. A workspace level
.vscode/settings file was added with the "eslint.runtime" key set to the
absolute path to the nvm installation of the correct Node version. Since
this path can vary, I decided to ignore the entire .vscode directory for
now. If there are ever any shareable .vscode configs, they can be added
later.